Rename new Storage.ts controller file
